+++ backend/kotlin-core/src/main/kotlin/br/com/zup/beagle/widget/core/UnitValue.kt
@@ -16,8 +16,6 @@
 
 package br.com.zup.beagle.widget.core
 
-import br.com.zup.beagle.ext.unitReal
-
 /**
  * Represents measurement values that contain both the numeric magnitude and the unit of measurement.
  * @property value the numeric measurement value.
